pyproject.toml
[project]
name = "video-processor"
version = "0.1.0"
description = "A video processing tool that extracts audio, transcribes using Nvidia Parakeet, generates titles/descriptions, and adds thumbnails"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "moviepy>=1.0.3",
    "nemo-toolkit[asr]>=1.23.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "python-slugify>=8.0.0",
    "pillow>=10.0.0",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
video-processor = "video_processor.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_processor"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
